use 128bit unsigned integers in util::pow_mod instead of 64bit unsigned integers to avoid overflow at k * k * n

+++ src/bellard.rs
@@ -19,7 +19,7 @@ pub fn pihex(d: u64) -> String {
     (0..4)
         .scan(fraction, |x, _| {
             *x = (*x - x.floor()) * 16.0;
-            Some(format!("{:x}", x.floor() as u64))
+            Some(format!("{:x}", x.floor() as u32))
         })
         .fold(String::new(), |s, t| s + &t)
 }
@@ -27,9 +27,9 @@ pub fn pihex(d: u64) -> String {
 fn series_sum(d: u64, j: u64, k: u64) -> f64 {
     let fraction1: f64 = (0..(2 * d + 2) / 5)
         .map(|i| {
-            ((if i % 2 == 0 { 1.0 } else { -1.0 }) *
-             util::pow_mod(4, 2 * d - 3 - 5 * i, j * i + k) as f64) /
-            ((j * i + k) as f64)
+            (if i % 2 == 0 { 1.0 } else { -1.0 }) *
+            util::pow_mod(4, (2 * d - 3 - 5 * i) as u128, (j * i + k) as u128) as f64 /
+            (j * i + k) as f64
         })
         .fold(0.0, |x, y| (x + y).fract());
     let fraction2: f64 = ((2 * d + 2) / 5..)
